Get dtype from patch emb

+++ vit_pytorch/mae.py
@@ -52,7 +52,7 @@ class MAE(nn.Module):
         if self.encoder.pool == "cls":
             tokens += self.encoder.pos_embedding[:, 1:(num_patches + 1)]
         elif self.encoder.pool == "mean":
-            tokens += self.encoder.pos_embedding.to(device) 
+            tokens += self.encoder.pos_embedding.to(device, dtype=tokens.dtype) 
 
         # calculate of patches needed to be masked, and get random indices, dividing it up for mask vs unmasked
